pyproject.toml
[tool.poetry]
name = "friend-chat"
version = "0.1.0"
description = ""
authors = ["aiyyskhan <aiyyskhan@gmail.com>"]
readme = "README.md"
packages = [{include = "friend_chat"}]

[tool.poetry.dependencies]
python = "^3.10"
openai = "^0.26.5"
numpy = "^1.24.2"
pytelegrambotapi = "^4.10.0"
python-dotenv = "^0.21.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
